+++ CHANGELOG.md
@@ -6,6 +6,7 @@ New minimum version requirement for the module is now:
 * azurerm 2.59.0
 
 - Remove hardcoded suffix from deployment name (#1)
+- `var.service_principal` should be an object (#2)
 - Default node pool should not be a list (#3)
 
 ## v1.0.0